requirements.txt
# Medical AI - Lung Disease Classification Requirements
# Version: 1.0.0
# Date: 2025-08-08

# Core ML and Deep Learning
tensorflow==2.19.0
keras==3.11.1
numpy==2.1.3
scikit-learn==1.7.1

# Data Processing and Visualization
pandas==2.3.1
matplotlib==3.10.5
pillow==11.3.0
seaborn==0.13.2

# Dataset Management
kagglehub==0.3.12
mlcroissant==1.0.21

# Web Framework and API
fastapi==0.116.1
uvicorn==0.35.0
python-multipart==0.0.20

# Database and ORM
motor==3.3.2
pymongo==4.6.1
beanie==1.25.0

# Additional Dependencies
scipy==1.16.1
h5py==3.14.0
joblib==1.5.1
python-dotenv==1.1.1 
passlib[bcrypt]==1.7.4
python-jose==3.3.0
bcrypt==4.1.3